fix(rabbitmq): close after consumer open failure
Discard failed native consumer handles before shutdown so they cannot be
closed through typed-nil interfaces. Align analyzers with Go 1.27 so the
regression reaches the broker-backed CI gate.

Impact: patch

+++ rabbitmq/adapter_contract_test.go
@@ -169,7 +169,7 @@ func TestAdapterDeadLettersMalformedAttemptMetadata(t *testing.T) {
 		rabbitmqqueue.Delivery{
 			Body: []byte("payload"), MessageID: "job-1",
 			Headers: []rabbitmqqueue.Header{
-				rabbitmqqueue.Int64Header(deliveryAttemptHeader, int64(job.MaxRetryCount+2)),
+				rabbitmqqueue.Int64Header(deliveryAttemptHeader, job.MaxRetryCount+2),
 			},
 		},
 		&message,
